Move component to separate file
